requirements.txt
Flask==2.2.2
flasgger==0.9.5
torch
transformers==4.31.0
Pillow==9.3.0
requests==2.31.0